requirements.txt
azure-ai-inference
Flask